# docs/static/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "B4 network packet processor REST API",
        "title": "B4 API",
        "contact": {},
        "version": "1.80.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/capture/upload": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Capture"
                ],
                "summary": "Upload a capture file",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Capture binary file",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Domain name",
                        "name": "domain",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "tls",
                        "description": "Protocol (tls or quic)",
                        "name": "protocol",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/geodat/upload": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Geodat"
                ],
                "summary": "Upload geodat file",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Geodat file (.dat or .db)",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "File type (geosite or geoip)",
                        "name": "type",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Destination directory path",
                        "name": "destination_path",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/system/update": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System"
                ],
                "summary": "Start update process",
                "parameters": [
                    {
                        "description": "Update request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.UpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.UpdateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.UpdateResponse"
                        }
                    }
                }
            }
        },
        "/system/update/upload": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Installs b4 from a b4-linux-\u003carch\u003e.tar.gz uploaded by the reader, for a\nnetwork where no download source can be reached. The archive is checked for\na b4 entry built for this machine before anything is replaced.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System"
                ],
                "summary": "Update from an uploaded release archive",
                "parameters": [
                    {
                        "type": "file",
                        "description": "b4-linux-\u003carch\u003e.tar.gz release archive",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Expected SHA256 of the archive",
                        "name": "sha256",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.UpdateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.UpdateResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handler.UpdateRequest": {
            "type": "object",
            "properties": {
                "version": {
                    "type": "string"
                }
            }
        },
        "handler.UpdateResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "service_manager": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                },
                "update_command": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter \"Bearer {token}\" to authorize",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}